Create Program 2a
Aim: Demonstrating creation of functions, passing parameters and return values
a) Defined as a function F as Fn = Fn-1 + Fn-2. Write a Python program which accepts a value for N (where N >0)
as input and pass this value to the function. Display suitable error message if the condition for input value is
not followed.